+++ src/collector/CollectorWrapper.ts
@@ -4,7 +4,7 @@ import { createSessionStartedUserAction } from '../user-action/createSessionStar
 import { CollectorOptions } from '../types'
 import ChangeEventListener from '../event-listeners/ChangeEventListener'
 import UserActionHandler from '../user-action/UserActionHandler'
-import UnloadEventListener from '../event-listeners/UnloadEventListener'
+import BeforeUnloadEventListener from '../event-listeners/BeforeUnloadEventListener'
 import KeyUpEventListener from '../event-listeners/KeyUpEventListener'
 import KeyDownEventListener from '../event-listeners/KeyDownEventListener'
 import { debugUserActionSessionSender, defaultUserActionSessionSender } from '../user-action/userActionSessionSender'
@@ -40,7 +40,7 @@ class CollectorWrapper {
     new KeyUpEventListener(this.userActionHandler, this.window).init()
     new KeyDownEventListener(this.userActionHandler, this.window).init()
     new ChangeEventListener(this.userActionHandler, this.window).init()
-    new UnloadEventListener(this.userActionHandler, this.window).init()
+    new BeforeUnloadEventListener(this.userActionHandler, this.window).init()
   }
 }