+++ src/event-listeners/BeforeUnloadEventListener.test.ts
@@ -2,10 +2,10 @@ import { beforeEach, describe, expect, it, vitest } from 'vitest'
 import UserActionHandler from '../user-action/UserActionHandler'
 import { waitFor } from '@testing-library/dom'
 import { nop } from '../utils/nop'
-import UnloadEventListener from '../event-listeners/UnloadEventListener'
+import BeforeUnloadEventListener from '../event-listeners/BeforeUnloadEventListener'
 import createElementInJSDOM from '../test-utils/createElementInJSDOM'
 
-describe('UnloadEventListener', () => {
+describe('BeforeUnloadEventListener', () => {
   describe('listener', () => {
     const userActionHandler = new UserActionHandler('aaa-111', 0, nop)
     const flushSpy = vitest.spyOn(userActionHandler, 'flush')
@@ -14,10 +14,10 @@ describe('UnloadEventListener', () => {
       vitest.restoreAllMocks()
     })
 
-    it('calls listener when unload event been fired', async () => {
+    it('calls listener when beforeunload event been fired', async () => {
       const { domWindow } = createElementInJSDOM('<div/>', 'div')
-      new UnloadEventListener(userActionHandler, domWindow).init()
-      domWindow.dispatchEvent(new Event('unload'))
+      new BeforeUnloadEventListener(userActionHandler, domWindow).init()
+      domWindow.dispatchEvent(new Event('beforeunload'))
 
       await waitFor(() => {
         expect(flushSpy).toHaveBeenCalledOnce()