+++ src/event-listeners/BeforeUnloadEventListener.ts
@@ -2,9 +2,9 @@ import UserActionHandler from '../user-action/UserActionHandler'
 import EventListener from '../event-listeners/EventListener'
 import { UserActionType } from '../types'
 
-class UnloadEventListener extends EventListener {
+class BeforeUnloadEventListener extends EventListener {
   constructor(userActionHandler: UserActionHandler, window: Window) {
-    super(userActionHandler, 'unload' as UserActionType, window)
+    super(userActionHandler, 'beforeunload' as UserActionType, window)
   }
 
   async listener(event: Event) {
@@ -12,4 +12,4 @@ class UnloadEventListener extends EventListener {
   }
 }
 
-export default UnloadEventListener
+export default BeforeUnloadEventListener